seperated all functional components inside checkout.jsx to seperate jsx files
